+++ src/models/pixel-office/index.ts
@@ -1,1 +1,2 @@
+export * from "./lobby-contracts";
 export * from "./scene-contracts";